+++ lib/Auth/userDetailsRegister.dart
@@ -121,7 +121,7 @@ class _UserDetailsRegisterState extends State<UserDetailsRegister> {
                             Navigator.push(
                               context,
                               MaterialPageRoute(
-                                  builder: (context) => Userhome()),
+                                  builder: (context) => Testhome2()),
                             );
                           }
                         },
